Move root stores into anchors.rs
